% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_plot.R
\name{accuracy_plot}
\alias{accuracy_plot}
\title{accuracy plot function takes in a tibble with mean of accuracy and neighbours and plots it}
\usage{
accuracy_plot(tibble)
}
\arguments{
\item{accuracies}{a tibble that contains accuracy estimates at different numbers of neighbours}
}
\value{
a plot that indicates accuracy estimates at different value of neighbours
}
\description{
accuracy plot function takes in a tibble with mean of accuracy and neighbours and plots it
}
\examples{
accuracy_plot(accuracy)
}
