% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_check.R
\name{balance_check}
\alias{balance_check}
\title{Creates a table}
\usage{
balance_check(df, col_n)
}
\arguments{
\item{df}{A dataframe or dataframe extension}

\item{col_n}{A quoted column name of the desigred column}
}
\value{
A table with three columns
}
\description{
take a data frame and the diagnosis column and creates a table that with the percentage for the diagnosis
must only have numeric types
}
\examples{
diagnosis_heart_disease(heart_disease_clean, "diagnosis")
}
