% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Cleans dataframe}
\usage{
clean_data(df)
}
\arguments{
\item{df}{A dataframe or dataframe extension}
}
\value{
A dataframe with only numeric data with no missing values
}
\description{
Finds all instances of missing data and converts it to NA
Then removes all rows with NA to create a new dataframe, this dataframe
must only have museric types
}
\examples{
clean_data(heart_disease)
}
