% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{diagnosis}
\alias{diagnosis}
\title{Cleans dataframe}
\usage{
diagnosis(df)
}
\arguments{
\item{df}{A dataframe or dataframe extension that must contain a column called diagnosis}
}
\value{
A dataframe with modified diagnosis input
}
\description{
take a data frame and change diagnosis column that is greater than 0 to 1
must only have numeric types
}
\examples{
diagnosis(heart_disease)
}
